Fix code review findings: type safety, dedup, null handling
- Change ExposedAirport.closest_quake_distance_km from int to float
- Remove duplicate ScoringMethod definition from scoring.py (import from config)
- Skip USGS events with null magnitude (unreviewed events)
- Consolidate daily-report pipeline steps so failures halt the job

+++ src/seismic_risk/models.py
@@ -68,7 +68,7 @@ class ExposedAirport:
     latitude: float
     longitude: float
     municipality: str
-    closest_quake_distance_km: int
+    closest_quake_distance_km: float
     nearby_quakes: list[NearbyQuake] = field(default_factory=list)
     exposure_score: float = 0.0
